requirements.txt
Django
psycopg2
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
drf_yasg
django-money
django-extensions
watchdog
pycountry